feat: scraper migration and v3 upgrade
