+++ apps/searchneu/package.json
@@ -8,15 +8,12 @@
     "start": "next start",
     "lint": "next lint",
     "test": "node src/scripts/run-tests.mjs",
-    "test:watch": "node --test --watch --import tsx src/**/*.test.ts",
-    "scrape:gen": "tsx ./src/scripts/scrape-gen.ts",
-    "scrape:up": "tsx ./src/scripts/scrape-up.ts",
-    "db:push": "drizzle-kit push",
-    "db:studio": "drizzle-kit studio"
+    "test:watch": "node --test --watch --import tsx src/**/*.test.ts"
   },
   "dependencies": {
-    "@sneu/db": "workspace:*",
     "@opentelemetry/api": "^1.9.0",
+    "@sneu/db": "workspace:*",
+    "@sneu/scraper": "workspace:*",
     "@tanstack/react-virtual": "^3.13.12",
     "@vercel/analytics": "^1.6.1",
     "@vercel/otel": "^2.1.0",
@@ -26,6 +23,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
+    "drizzle-orm": "^0.45.0",
     "flags": "^4.0.2",
     "input-otp": "^1.4.2",
     "jose": "^6.1.3",
@@ -46,23 +44,17 @@
     "@next/env": "^16.0.7",
     "@tailwindcss/postcss": "^4.1.17",
     "@tailwindcss/typography": "^0.5.19",
-    "@types/he": "^1.2.3",
-    "@types/mdx": "^2.0.13",
     "@types/node": "^24.10.1",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "babel-plugin-react-compiler": "^1.0.0",
     "eslint": "^9.39.1",
     "eslint-config-next": "^16.0.7",
-    "he": "^1.2.0",
-    "node-html-parser": "^7.0.1",
     "prettier": "^3.7.4",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "server-only": "^0.0.1",
     "tailwindcss": "^4.1.17",
-    "tsx": "^4.21.0",
-    "typescript": "^5.9.3",
-    "yaml": "^2.8.2"
+    "typescript": "^5.9.3"
   },
   "prettier": {
     "plugins": [
