+++ package.json
@@ -12,7 +12,7 @@
   "devDependencies": {
     "turbo": "^2.6.3"
   },
-  "packageManager": "pnpm@10.24.0",
+  "packageManager": "pnpm@10.25.0",
   "engine": {
     "node": "24.x"
   }
