+++ packages/scraper/src/schemas/sectionCatalogDetails.ts
@@ -1,0 +1,3 @@
+import * as z from "zod";
+
+export const BannerSectionCatalogDetails = z.string();
