+++ packages/scraper/src/schemas/sectionDescription.ts
@@ -1,0 +1,3 @@
+import * as z from "zod";
+
+export const BannerSectionDescription = z.string();
