+++ packages/scraper/src/schemas/sectionPrereqs.ts
@@ -1,0 +1,3 @@
+import * as z from "zod";
+
+export const BannerSectionPrereqs = z.string();
